+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Print warning in logs if log verbosity is set to trace
 
+### Fixed
+- Parallel sessions may fail when one of the sessions requires pairing
+
 ## [0.12.4] - 2023-05-16
 
 ### Fixed