+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- Print warning in logs if log verbosity is set to trace
+
 ### Fixed
 - LogoPath is incorrect after a requestor scheme update